backend/requirements.txt
agno
python-dotenv
fastapi
uvicorn
google-genai
pymongo
RestrictedPython